adding nouvelle carte

+++ src/Blockchain.java
@@ -1,8 +1,10 @@
 package src;
 
 import java.io.FileWriter;
+import java.io.File;
 import java.io.IOException;
 import java.io.PrintWriter;
+import java.nio.file.Files;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -204,27 +206,54 @@ public class Blockchain{
 
     public void writeBlockchain() {
         try {
-            FileWriter fileWriter = new FileWriter("blockchain.txt");
-            PrintWriter printWriter = new PrintWriter(fileWriter);
-            for(int i = 0; i < this.blocks.size(); i++){
-                printWriter.printf("%d\n",this.blocks.get(i).getIndex());
-                printWriter.println(this.blocks.get(i).getTimestamp().toString());
-                printWriter.println(this.blocks.get(i).getPreviousHash());
-                printWriter.println(this.blocks.get(i).getCurrentHash());
-                printWriter.println(this.blocks.get(i).getStudent().getIdStudent());
-                printWriter.println(this.blocks.get(i).getStudent().getNom());
-                printWriter.println(this.blocks.get(i).getStudent().getPrenom());
-                printWriter.println(this.blocks.get(i).getStudent().getHashID());
-                printWriter.println(this.blocks.get(i).getStudent().getHashJAPD());
-                printWriter.println(this.blocks.get(i).getStudent().getHashBAC());
-                for(int j = 0; j < this.blocks.get(i).getStudent().getDiplomes().size(); j++)
-                    printWriter.println(Arrays.toString(this.blocks.get(i).getStudent().getDiplomes().get(j)));
+            File tempFile = new File("blockchain.txt");
+            boolean exists = tempFile.exists();
 
-                printWriter.println();
+            if(!exists) {
+                FileWriter fileWriter = new FileWriter("blockchain.txt");
+                PrintWriter printWriter = new PrintWriter(fileWriter);
+                for (int i = 0; i < this.blocks.size(); i++) {
+                    printWriter.printf("%d\n", this.blocks.get(i).getIndex());
+                    printWriter.println(this.blocks.get(i).getTimestamp().toString());
+                    printWriter.println(this.blocks.get(i).getPreviousHash());
+                    printWriter.println(this.blocks.get(i).getCurrentHash());
+                    printWriter.println(this.blocks.get(i).getStudent().getIdStudent());
+                    printWriter.println(this.blocks.get(i).getStudent().getNom());
+                    printWriter.println(this.blocks.get(i).getStudent().getPrenom());
+                    printWriter.println(this.blocks.get(i).getStudent().getHashID());
+                    printWriter.println(this.blocks.get(i).getStudent().getHashJAPD());
+                    printWriter.println(this.blocks.get(i).getStudent().getHashBAC());
+                    for (int j = 0; j < this.blocks.get(i).getStudent().getDiplomes().size(); j++)
+                        printWriter.println(Arrays.toString(this.blocks.get(i).getStudent().getDiplomes().get(j)));
+
+                    printWriter.println("DONE");
+                }
+                printWriter.close();
+            }else{
+                FileWriter fileWriter = new FileWriter("blockchain.txt", true);
+                PrintWriter printWriter = new PrintWriter(fileWriter);
+                for (int i = 0; i < this.blocks.size(); i++) {
+                    printWriter.printf("%d\n", this.blocks.get(i).getIndex());
+                    printWriter.println(this.blocks.get(i).getTimestamp().toString());
+                    printWriter.println(this.blocks.get(i).getPreviousHash());
+                    printWriter.println(this.blocks.get(i).getCurrentHash());
+                    printWriter.println(this.blocks.get(i).getStudent().getIdStudent());
+                    printWriter.println(this.blocks.get(i).getStudent().getNom());
+                    printWriter.println(this.blocks.get(i).getStudent().getPrenom());
+                    printWriter.println(this.blocks.get(i).getStudent().getHashID());
+                    printWriter.println(this.blocks.get(i).getStudent().getHashJAPD());
+                    printWriter.println(this.blocks.get(i).getStudent().getHashBAC());
+                    for (int j = 0; j < this.blocks.get(i).getStudent().getDiplomes().size(); j++)
+                        printWriter.println(Arrays.toString(this.blocks.get(i).getStudent().getDiplomes().get(j)));
+
+                    printWriter.println("DONE");
+                }
+                printWriter.close();
             }
-            printWriter.close();
         }catch(IOException e){
             e.printStackTrace();
         }
     }
+
+
 }
